+++ middlewares/cors.js
@@ -1,0 +1,7 @@
+const cors = require('cors');
+
+module.exports = function(req, res, next) {
+    cors();
+    next();
+};
+
