+++ middlewares/noCache.js
@@ -1,0 +1,4 @@
+module.exports = function (req, res, next) {
+    res.header('Cache-Control', 'no-cache');
+    next()
+};
