+++ package.json
@@ -2,10 +2,10 @@
   "name": "nodeinitialdemo",
   "version": "1.0.0",
   "description": "Victor's fork of nodeInitialDemo",
-  "main": "app/app.js",
+  "main": "app.js",
   "scripts": {
-    "start": "node app/app.js",
-    "dev": "nodemon app/app.js",
+    "start": "node app.js",
+    "dev": "nodemon app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Victor Machado Perez",
